+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Revision history for hkd-records
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.0.1 -- 2022-07-22
 
 * First version. Released on an unsuspecting world.
+
+## 0.0.2 -- 2022-07-23
+
+* added documentation
+* changed generic functions to default member functions
+* changed HkdProd and LkdProd type names
+